refactor(cli): establish typed application boundaries
Compile the CLI through one library-owned entrypoint so the binary no longer duplicates the full module graph and unit-test suite.

Remove unfinished context and layout surfaces, make file-history diffs exact and failure-visible, and route model-catalog recovery through typed HTTP/transport errors instead of rendered text. Add behavior tests for literal URL text, checkpoint failures, and token refresh/retry, plus a first-principles follow-up plan.
