+++ Cargo.toml
@@ -79,6 +79,7 @@ serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 serde_yaml_ng = "0.10.0"
 sha2 = "0.10.9"
+similar = "2.7.0"
 hex = "0.4.3"
 sqlx = { version = "0.8.6", default-features = false, features = ["mysql", "runtime-tokio-rustls"] }
 tokio = { version = "1.47.1", features = ["macros", "net", "process", "rt-multi-thread", "signal"] }
