+++ crates/astra-cli/Cargo.toml
@@ -58,6 +58,7 @@ serde.workspace = true
 serde_json.workspace = true
 serde_yaml_ng.workspace = true
 sha2.workspace = true
+similar.workspace = true
 base64.workspace = true
 sqlx.workspace = true
 supports-color = "3"
