+++ crates/astra-cli/src/lib.rs
@@ -31,6 +31,7 @@ pub mod admin_cli;
 pub(crate) mod background_task_error;
 pub mod diff_utils;
 pub mod edge_tools;
+pub mod entrypoint;
 pub mod explain_dag;
 pub mod git_branch_cache;
 pub mod lock_recovery;
@@ -73,4 +74,53 @@ pub(crate) mod tests {
     pub(crate) use super::test_utils::stub_stream_result_with_records;
     pub(crate) use super::test_utils::test_temp_dir;
     pub(crate) use super::test_utils::wait_until;
+
+    pub(crate) use crate::cli::slash::slash_session::resolve_journal_target_session;
+    pub(crate) use crate::cli::slash::slash_task;
+    pub(crate) use astra_services::session_journal;
+    use axum::{Router, routing::get};
+
+    async fn mock_models_response() -> axum::Json<serde_json::Value> {
+        axum::Json(serde_json::json!({
+            "models": [
+                {
+                    "name": "test-model",
+                    "is_active": true,
+                    "context_window": 200_000
+                },
+                {
+                    "name": "mock-model",
+                    "is_active": true,
+                    "context_window": 200_000
+                }
+            ]
+        }))
+    }
+
+    async fn spawn_mock_app(app: Router) -> String {
+        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
+        let addr = listener.local_addr().unwrap();
+        let base = format!("http://{addr}");
+        tokio::spawn(async move {
+            axum::serve(listener, app).await.ok();
+        });
+        tokio::task::yield_now().await;
+        base
+    }
+
+    async fn spawn_mock(app: Router) -> String {
+        spawn_mock_app(app.route("/models", get(mock_models_response))).await
+    }
+
+    mod auth_tests;
+    mod chat_stream_tests;
+    mod chat_turn_tests;
+    mod cli_args_tests;
+    mod cloud_sync_tests;
+    mod cost_tracking_tests;
+    mod preamble_tests;
+    mod resume_tests;
+    mod self_command_tests;
+    mod slash_command_tests;
+    mod stats_tools_tests;
 }
