--- crates/astra-cli/src/tui/layout.rs
@@ -1,2 +1,0 @@
-// Dynamic layout is computed inline in mod.rs using bottom_pane.desired_height(width).
-// This module is reserved for future layout utilities if needed.
